# Speed up fish completions for subcommands