The current method of generating completions for the fish shell relies
on running `docker-compose --help`. On my system (a 2018 MacBook Pro)
this routinely takes several hundred milliseconds:

```
$ hyperfine --warmup 3 'docker-compose --help'
Benchmark #1: docker-compose --help
  Time (mean ± σ):     351.6 ms ±  16.0 ms    [User: 295.5 ms, System: 44.3 ms]
  Range (min … max):   332.8 ms … 384.0 ms    10 runs
```

If I configure fish to automatically load these completions, it makes
launching a new fish process very sluggish.

In addition, the regular expression used to parse the output of
`docker-compose --help` incorrectly generates completions for `name` and
`in` due to the following lines:

https://github.com/docker/compose/blob/9c5351cf27d10a7b8078496fb21a37929b6687cb/compose/cli/main.py#L208
https://github.com/docker/compose/blob/9c5351cf27d10a7b8078496fb21a37929b6687cb/compose/cli/main.py#L212

This commit fixes these issues by explicitly listing subcommands instead
of generating them dynamically. This means that adding/removing
subcommands won't automatically update the completions but according to
the git blame for `compose/cli/main.py`, no one has touched the list of
subcommands in the help message in 3+ years so I hope that won't be an
issue.

<!--
Welcome to the docker-compose issue tracker, and thank you for your interest
in contributing to the project! Please make sure you've read the guidelines
in CONTRIBUTING.md before submitting your pull request. Contributions that
do not comply and contributions with failing tests will not be reviewed!
-->

<!-- Please make sure an issue describing the problem the PR is trying to
    solve exists, or create it before submitting a PR. The maintainers will
    validate if the issue should be addressed or if it is out of scope for the
    project.
-->
Resolves #7418 

Update: I discovered that fish has a variable `fish_complete_path` which determines which directories it will search for completions in. By adding the vendor completions directory to that variable, I can make fish defer loading the completion until it's actually needed (i.e., only after typing `docker-compose <TAB>`). This speeds up launching new fish processes but there's still the issue that when I _do_ want to load the completions, it hangs for a good while before displaying them (not to mention the incorrect completions bug). So I still think this PR is helpful.