@ulyssessouza 2020-05-20
Hello @cstyles ,

Well, for the wrong completions, I think that the path is to fix the regexp. But changing the approach to a fixed list I don't agree, since it will increase the time in maintainability and it will be for sure something to forget.