@cstyles 2020-05-30
I agree that, at least in theory, it's not the cleanest solution. But in practice, the list of subcommands hasn't changed in 3 years so I don't think maintenance will take up that much time (especially since fish isn't exactly the most popular shell). On top of that, the completions for bash and zsh both hardcode the subcommands as well (plus all their options) so this approach is at least consistent.

IMO the best solution is to improve the performance of `docker-compose --help` but that would probably require a lot of effort. I figured just hardcoding the completions gets most of the benefits with little cost. Just my two cents.